pyproject.toml
[project]
name = "palabras-perdidas"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai>=2.3.0",
    "python-dotenv>=1.1.1",
    "tqdm>=4.67.1",
]
